requirements.txt
aiogram
python-dotenv
requests
transformers
torch
pandas
jupyter
scikit-learn
openai
matplotlib
seaborn
datasets
